requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
click==7.1.2
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.5
django-bootstrap-themes==3.3.6
django-debug-toolbar==2.2
django-heroku==0.3.1
django-import-export==2.1.0
django-liststyle==0.2b0
django-mathfilters==1.0.0
dnspython==1.16.0
et-xmlfile==1.0.1
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
importlib-metadata==1.6.0
jdcal==1.4.1
kombu==4.6.8
MarkupPy==1.14
mysql-connector-python==8.0.19
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==3.0.3
pkg-resources==0.0.0
protobuf==3.6.1
psycopg2==2.8.5
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
uvicorn==0.11.5
uvloop==0.14.0
vine==1.3.0
websockets==8.1
whitenoise==5.1.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
